Work around a problem with strange synthesised ASTs
For some reason, the match expression in Tox{Av,Core}Impl.iterate was
turned into a piece of code like this:

      case <synthetic> val x1: Option[im.tox.tox4j.av.proto.Av.AvEvents] = x;
      case6(){
        // Oh noes!
        if (scala.None.==(x1))
          matchEnd5(())
        else
          case7()
      };

Our None== checker would flag this as an error. I don't currently see
how we can detect this. Checking whether the (x1) expression is
synthetic causes many false negatives.

I don't even know how to reproduce this. Every attempt in the test suite
fails to produce this tree. I don't know what interactions may be
causing this.

Thus, the solution for now is to simply avoid the problem by using
Option.foreach instead of the match expression. In the process of trying
to debug the problem, I split up the large .iterate() methods into many
smaller ones. I'm keeping it this way.