Add some more tests for bad finalize order.
This code should never be executed during normal execution. We guard
against using an instance after it has been garbage collected, which is
not normally possible.

However, using reflection, users can modify the final field in a Tox
instance to create invalid instances. We don't want this to crash the
JVM, so instead we throw exceptions.